ISSUE-3287: Move AddOnStream to transforms/streams
