+++ query/src/interpreters/mod.rs
@@ -45,7 +45,6 @@ mod interpreter_user_alter;
 mod interpreter_user_create;
 mod interpreter_user_drop;
 mod plan_schedulers;
-mod stream_addon;
 
 pub use interpreter::Interpreter;
 pub use interpreter::InterpreterPtr;
@@ -79,4 +78,3 @@ pub use interpreter_user_alter::AlterUserInterpreter;
 pub use interpreter_user_create::CreateUserInterpreter;
 pub use interpreter_user_drop::DropUserInterpreter;
 pub use plan_schedulers::PlanScheduler;
-pub use stream_addon::AddOnStream;
