+++ query/src/pipelines/transforms/mod.rs
@@ -12,6 +12,27 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod transform_aggregator_final;
+mod transform_aggregator_partial;
+mod transform_create_sets;
+mod transform_expression;
+mod transform_expression_executor;
+mod transform_filter;
+mod transform_group_by_final;
+mod transform_group_by_partial;
+mod transform_limit;
+mod transform_limit_by;
+mod transform_projection;
+mod transform_remote;
+mod transform_sort_merge;
+mod transform_sort_partial;
+mod transform_source;
+
+mod group_by;
+mod streams;
+mod transform_sink;
+
+pub use streams::AddOnStream;
 pub use transform_aggregator_final::AggregatorFinalTransform;
 pub use transform_aggregator_partial::AggregatorPartialTransform;
 pub use transform_create_sets::CreateSetsTransform;
@@ -31,22 +52,3 @@ pub use transform_sort_merge::SortMergeTransform;
 pub use transform_sort_partial::get_sort_descriptions;
 pub use transform_sort_partial::SortPartialTransform;
 pub use transform_source::SourceTransform;
-
-mod transform_aggregator_final;
-mod transform_aggregator_partial;
-mod transform_create_sets;
-mod transform_expression;
-mod transform_expression_executor;
-mod transform_filter;
-mod transform_group_by_final;
-mod transform_group_by_partial;
-mod transform_limit;
-mod transform_limit_by;
-mod transform_projection;
-mod transform_remote;
-mod transform_sort_merge;
-mod transform_sort_partial;
-mod transform_source;
-
-mod group_by;
-mod transform_sink;
