+++ query/src/pipelines/transforms/streams/stream_addon.rs
